feat(displayChatRoom.js): Import updatePlayers and send leave room event
Import updatePlayers which is a replacement for the "all players"
event so it can be reused. The leave room btn now sends "leave room" event
and it also updates the players displayed. The updatePlayers is also
invoked when displayChatRoom is invoked since it needs
to updatePlayers everytime someone enters as well.