feat(updatePlayers.js): add function to update players in room
This function will listen to "all players" event that will
be emitted when someone emits "leave room" or "join room".
The arg that is sent back will be an array of all users
connected to that specific room. And the cb function contains
code for updating the elements in DOM.